third_party/blink/web_tests/wpt_internal/css/css-transitions/initial-pseudo-parsing.html

<!DOCTYPE html>
<title>CSS Transitions: Parsing tests for :initial pseudo class</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8174">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  // Valid selectors that will match.
  test_valid_selector(":initial");
  test_valid_selector("div:initial#id");
  test_valid_selector("div + :initial");

  // Valid selectors that parse, but will never match.
  test_valid_selector("div:initial span");
  test_valid_selector(":is(:initial)");
  test_valid_selector(":where(:hover, :initial)");
  test_valid_selector(":is(:is(div, :initial), :has(:initial))");
  test_valid_selector(":nth-child(2n+1 of :initial)");

  // Typographic pseudo elements.
  test_invalid_selector("::first-line:initial");
  test_invalid_selector("::first-letter:initial");
  test_invalid_selector("::prefix:initial");
  test_invalid_selector("::postfix:initial");

  // Highlight pseudo elements.
  test_invalid_selector("::selection:initial");
  test_invalid_selector("::target-text:initial");
  test_invalid_selector("::spelling-error:initial");
  test_invalid_selector("::grammar-error:initial");

  // Tree-abiding pseudo elements
  test_valid_selector("::before:initial");
  test_valid_selector("::after:initial");
  test_valid_selector("::marker:initial");
  test_valid_selector("::placeholder:initial");
  test_valid_selector("::file-selector-button:initial");

  // Shadow pseudo elements.
  test_valid_selector("::part(part):initial");
  test_valid_selector("::slotted(:initial)");
  test_invalid_selector("::slotted(*):initial");
</script>
